added general StatesFactory interface, moved factory implementation to LazyStatesFactory
