+++ src/client/client.cpp
@@ -5,7 +5,7 @@
 #include "TaskManagerService.grpc.pb.h"
 
 #include "view/commands/factory/CommandFactory.h"
-#include "view/states/factory/StatesFactory.h"
+#include "view/states/factory/LazyStatesFactory.h"
 #include "view/user_interface/console_io/ConsolePrinter.h"
 #include "view/user_interface/console_io/ConsoleReader.h"
 #include "view/user_interface/UserInterface.h"
@@ -20,20 +20,16 @@
 int main() {
     std::string target_str = "localhost:8586";
 
-    std::shared_ptr<CommandFactory> command_factory = std::make_shared<CommandFactory>();
-    std::shared_ptr<ConsolePrinter> printer = std::make_shared<ConsolePrinter>();
-    std::shared_ptr<ConsoleReader> reader = std::make_shared<ConsoleReader>();
-    std::shared_ptr<StatesFactory> states_factory = std::make_shared<StatesFactory>(command_factory,
-                                                                                    printer,
-                                                                                    reader);
-    std::unique_ptr<UserInterface> user_interface = std::make_unique<UserInterface>(states_factory, printer);
-    std::unique_ptr<ModelController> model_controller = std::make_unique<DefaultModelController>(
+    auto command_factory = std::make_shared<CommandFactory>();
+    auto printer = std::make_shared<ConsolePrinter>();
+    auto reader = std::make_shared<ConsoleReader>();
+    auto states_factory = std::make_shared<LazyStatesFactory>(command_factory, printer, reader);
+    auto user_interface = std::make_unique<UserInterface>(states_factory, printer);
+    auto model_controller = std::make_unique<DefaultModelController>(
             std::make_unique<GRPCClientEndPoint>(TaskManagerService::NewStub(grpc::CreateChannel(target_str, grpc::InsecureChannelCredentials()))),
             std::make_unique<TaskValidator>(),
             std::make_unique<persistence::PersistenceFactory>());
-    std::unique_ptr<ViewController> view_controller = std::make_unique<ViewController>(
-            std::move(model_controller),
-            std::move(user_interface));
+    auto view_controller = std::make_unique<ViewController>(std::move(model_controller), std::move(user_interface));
 
     view_controller->RunUserInterface();
 }
